Move solutions to buried package: project 05
